# Optimize string building for NamespacedName